+++ storage/fcds/leveldb/fcds_test.go
@@ -14,48 +14,39 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with the Swarm library. If not, see <http://www.gnu.org/licenses/>.
 
-package leveldb
+package fcds
 
 import (
 	"fmt"
 	"io/ioutil"
 	"math/rand"
 	"os"
-	"path/filepath"
 	"sync"
 	"testing"
 	"time"
 
 	"github.com/ethersphere/swarm/chunk"
-	"github.com/ethersphere/swarm/storage/fcds"
 )
 
 const (
 	ConcurrentThreads = 128
 )
 
-// NewFCDSStore is a test helper function that constructs
-// a new Store for testing purposes into which a specific MetaStore can be injected.
-func NewFCDSStore(t *testing.B) (s *fcds.Store, clean func()) {
-	t.Helper()
+func newDB(b *testing.B) (db Storer, clean func()) {
+	b.Helper()
 
-	path, err := ioutil.TempDir("", "swarm-fcds-")
+	path, err := ioutil.TempDir("/tmp/swarm", "swarm-shed")
 	if err != nil {
-		t.Fatal(err)
+		b.Fatal(err)
 	}
 
-	metaStore, err := NewMetaStore(filepath.Join(path, "meta"))
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	s, err = fcds.New(path, chunk.DefaultSize, metaStore)
+	db, err = New(path)
 	if err != nil {
 		os.RemoveAll(path)
-		t.Fatal(err)
+		b.Fatal(err)
 	}
-	return s, func() {
-		s.Close()
+	return db, func() {
+		db.Close()
 		os.RemoveAll(path)
 	}
 }
@@ -90,189 +81,215 @@ func GenerateTestRandomChunk() chunk.Chunk {
 	return chunk.NewChunk(key, data)
 }
 
-// Benchmarkings
-func runBenchmark(b *testing.B, baseChunksCount int, writeChunksCount int, readChunksCount int, deleteChunksCount int, iterationCount int) {
-	b.Helper()
+func createBenchBaseline(b *testing.B, baseChunksCount int) (db Storer, clean func(), baseChunks []chunk.Chunk) {
+	db, clean = newDB(b)
 
+	if baseChunksCount > 0 {
+		baseChunks = getChunks(baseChunksCount, baseChunks)
+		//start := time.Now()
+		sem := make(chan struct{}, ConcurrentThreads)
+		var wg sync.WaitGroup
+		wg.Add(baseChunksCount)
+		for i, ch := range baseChunks {
+			sem <- struct{}{}
+			go func(i int, ch chunk.Chunk) {
+				defer func() {
+					<-sem
+					wg.Done()
+				}()
+				if err := db.Put(ch); err != nil {
+					panic(err)
+				}
+			}(i, ch)
+		}
+		wg.Wait()
+		//elapsed := time.Since(start)
+		//fmt.Println("-- adding base chunks took, ", elapsed)
+	}
+
+	rand.Shuffle(baseChunksCount, func(i, j int) {
+		baseChunks[i], baseChunks[j] = baseChunks[j], baseChunks[i]
+	})
+
+	return db, clean, baseChunks
+}
+
+// Benchmarkings
+
+func runBenchmark(b *testing.B, db Storer, basechunks []chunk.Chunk, baseChunksCount int, writeChunksCount int, readChunksCount int, deleteChunksCount int) {
 	var writeElapsed time.Duration
 	var readElapsed time.Duration
 	var deleteElapsed time.Duration
 
-	for i := 0; i < iterationCount; i++ {
-		db, clean := NewFCDSStore(b)
-		var basechunks []chunk.Chunk
+	var writeChunks []chunk.Chunk
+	writeChunks = getChunks(writeChunksCount, writeChunks)
+	b.StartTimer()
 
-		b.StopTimer()
-		if baseChunksCount > 0 {
-			basechunks = getChunks(baseChunksCount, basechunks)
+	var jobWg sync.WaitGroup
+	if writeChunksCount > 0 {
+		jobWg.Add(1)
+		go func() {
 			start := time.Now()
 			sem := make(chan struct{}, ConcurrentThreads)
 			var wg sync.WaitGroup
-			wg.Add(baseChunksCount)
-			for i, ch := range basechunks {
+			wg.Add(writeChunksCount)
+			for i, ch := range writeChunks {
 				sem <- struct{}{}
 				go func(i int, ch chunk.Chunk) {
 					defer func() {
 						<-sem
 						wg.Done()
 					}()
-					if _, err := db.Put(ch); err != nil {
+					if err := db.Put(ch); err != nil {
 						panic(err)
 					}
 				}(i, ch)
 			}
 			wg.Wait()
 			elapsed := time.Since(start)
-			fmt.Println("-- adding base chunks took, ", elapsed)
-		}
+			fmt.Println("-- writing chunks took , ", elapsed)
+			writeElapsed += elapsed
+			jobWg.Done()
+		}()
+	}
 
-		rand.Shuffle(baseChunksCount, func(i, j int) {
-			basechunks[i], basechunks[j] = basechunks[j], basechunks[i]
+	if readChunksCount > 0 {
+		jobWg.Add(1)
+		go func() {
+			errCount := 0
+			start := time.Now()
+			sem := make(chan struct{}, ConcurrentThreads*4)
+			var wg sync.WaitGroup
+			wg.Add(readChunksCount)
+			for i, ch := range basechunks {
+				if i >= readChunksCount {
+					break
+				}
+				sem <- struct{}{}
+				go func(i int, ch chunk.Chunk) {
+					defer func() {
+						<-sem
+						wg.Done()
+					}()
+					_, err := db.Get(ch.Address())
+					if err != nil {
+						//panic(err)
+						errCount++
+					}
+				}(i, ch)
+			}
+			wg.Wait()
+			elapsed := time.Since(start)
+			//fmt.Println("-- reading chunks took , ", elapsed)
+			readElapsed += elapsed
+			jobWg.Done()
+		}()
+	}
+
+	if deleteChunksCount > 0 {
+		jobWg.Add(1)
+		go func() {
+			start := time.Now()
+			sem := make(chan struct{}, ConcurrentThreads)
+			var wg sync.WaitGroup
+			wg.Add(deleteChunksCount)
+			for i, ch := range basechunks {
+				if i >= deleteChunksCount {
+					break
+				}
+				sem <- struct{}{}
+				go func(i int, ch chunk.Chunk) {
+					defer func() {
+						<-sem
+						wg.Done()
+					}()
+					if err := db.Delete(ch.Address()); err != nil {
+						panic(err)
+					}
+				}(i, ch)
+			}
+			wg.Wait()
+			elapsed := time.Since(start)
+			//fmt.Println("-- deleting chunks took , ", elapsed)
+			deleteElapsed += elapsed
+			jobWg.Done()
+		}()
+	}
+
+	jobWg.Wait()
+
+	//if writeElapsed > 0 {
+	//	fmt.Println("- Average write  time : ", writeElapsed.Nanoseconds()/int64(iterationCount), " ns/op")
+	//}
+	//if readElapsed > 0 {
+	//	//fmt.Println("- Average read time : ", readElapsed.Nanoseconds()/int64(iterationCount), " ns/op")
+	//}
+	//if deleteElapsed > 0 {
+	//	//fmt.Println("- Average delete time : ", deleteElapsed.Nanoseconds()/int64(iterationCount), " ns/op")
+	//}
+}
+
+func BenchmarkWrite_Add10K(b *testing.B) {
+	for i := 10000; i <= 1000000; i *= 10 {
+		b.Run(fmt.Sprintf("Baseline_%d", i), func(b *testing.B) {
+			for j := 0; j < b.N; j++ {
+				b.StopTimer()
+				db, clean, baseChunks := createBenchBaseline(b, i)
+				b.StartTimer()
+
+				runBenchmark(b, db, baseChunks, 0, 10000, 0, 0)
+				b.StopTimer()
+				clean()
+				b.StartTimer()
+			}
 		})
-		b.StartTimer()
-
-		var jobWg sync.WaitGroup
-		if writeChunksCount > 0 {
-			jobWg.Add(1)
-			go func() {
-				var writeChunks []chunk.Chunk
-				writeChunks = getChunks(writeChunksCount, writeChunks)
-				start := time.Now()
-				sem := make(chan struct{}, ConcurrentThreads)
-				var wg sync.WaitGroup
-				wg.Add(writeChunksCount)
-				for i, ch := range writeChunks {
-					sem <- struct{}{}
-					go func(i int, ch chunk.Chunk) {
-						defer func() {
-							<-sem
-							wg.Done()
-						}()
-						if _, err := db.Put(ch); err != nil {
-							panic(err)
-						}
-					}(i, ch)
-				}
-				wg.Wait()
-				elapsed := time.Since(start)
-				fmt.Println("-- writing chunks took , ", elapsed)
-				writeElapsed += elapsed
-				jobWg.Done()
-			}()
-		}
-
-		if readChunksCount > 0 {
-			jobWg.Add(1)
-			go func() {
-				errCount := 0
-				start := time.Now()
-				sem := make(chan struct{}, ConcurrentThreads*4)
-				var wg sync.WaitGroup
-				wg.Add(readChunksCount)
-				for i, ch := range basechunks {
-					if i >= readChunksCount {
-						break
-					}
-					sem <- struct{}{}
-					go func(i int, ch chunk.Chunk) {
-						defer func() {
-							<-sem
-							wg.Done()
-						}()
-						_, err := db.Get(ch.Address())
-						if err != nil {
-							//panic(err)
-							errCount++
-						}
-					}(i, ch)
-				}
-				wg.Wait()
-				elapsed := time.Since(start)
-				fmt.Println("-- reading chunks took , ", elapsed)
-				readElapsed += elapsed
-				jobWg.Done()
-			}()
-		}
-
-		if deleteChunksCount > 0 {
-			jobWg.Add(1)
-			go func() {
-				start := time.Now()
-				sem := make(chan struct{}, ConcurrentThreads)
-				var wg sync.WaitGroup
-				wg.Add(deleteChunksCount)
-				for i, ch := range basechunks {
-					if i >= deleteChunksCount {
-						break
-					}
-					sem <- struct{}{}
-					go func(i int, ch chunk.Chunk) {
-						defer func() {
-							<-sem
-							wg.Done()
-						}()
-						if err := db.Delete(ch.Address()); err != nil {
-							panic(err)
-						}
-					}(i, ch)
-				}
-				wg.Wait()
-				elapsed := time.Since(start)
-				fmt.Println("-- deleting chunks took , ", elapsed)
-				deleteElapsed += elapsed
-				jobWg.Done()
-			}()
-		}
-
-		jobWg.Wait()
-		clean()
-	}
-
-	if writeElapsed > 0 {
-		fmt.Println("- Average write  time : ", writeElapsed.Nanoseconds()/int64(iterationCount), " ns/op")
-	}
-	if readElapsed > 0 {
-		fmt.Println("- Average read time : ", readElapsed.Nanoseconds()/int64(iterationCount), " ns/op")
-	}
-	if deleteElapsed > 0 {
-		fmt.Println("- Average delete time : ", deleteElapsed.Nanoseconds()/int64(iterationCount), " ns/op")
 	}
 }
 
-//func TestStorage (b *testing.T) {
-//  runBenchmark(b, 0, 1000000, 0, 0)
+func BenchmarkReadOverClean(b *testing.B) {
+	for i := 10000; i <= 1000000; i *= 10 {
+		b.Run(fmt.Sprintf("Baseline_%d", i), func(b *testing.B) {
+			for j := 0; j < b.N; j++ {
+				b.StopTimer()
+				db, clean, baseChunks := createBenchBaseline(b, i)
+				b.StartTimer()
+
+				runBenchmark(b, db, baseChunks, 0, 0, 10000, 0)
+				b.StopTimer()
+				clean()
+				b.StartTimer()
+			}
+		})
+	}
+}
+
+//func BenchmarkWriteOverClean_100000(t *testing.B) { runBenchmark(t, 0, 100000, 0, 0, 6) }
+//func BenchmarkWriteOverClean_1000000(t *testing.B) { runBenchmark(t, 0, 1000000, 0, 0, 4) }
+
+//func BenchmarkWriteOver1Million_10000(t *testing.B) {
+//for i := 0; i < t.N; i++ {
+//runBenchmark(t, 1000000, 10000, 0, 0, 8)
 //}
 
-func BenchmarkWriteOverClean_10000(t *testing.B)   { runBenchmark(t, 0, 10000, 0, 0, 8) }
-func BenchmarkWriteOverClean_100000(t *testing.B)  { runBenchmark(t, 0, 100000, 0, 0, 6) }
-func BenchmarkWriteOverClean_1000000(t *testing.B) { runBenchmark(t, 0, 1000000, 0, 0, 4) }
+//}
 
-func BenchmarkWriteOver1Million_10000(t *testing.B)   { runBenchmark(t, 1000000, 10000, 0, 0, 8) }
-func BenchmarkWriteOver1Million_100000(t *testing.B)  { runBenchmark(t, 1000000, 100000, 0, 0, 6) }
-func BenchmarkWriteOver1Million_1000000(t *testing.B) { runBenchmark(t, 1000000, 1000000, 0, 0, 4) }
+//func BenchmarkWriteOver1Million_100000(t *testing.B) { runBenchmark(t, 1000000, 100000, 0, 0,6) }
+//func BenchmarkWriteOver1Million_1000000(t *testing.B) { runBenchmark(t, 1000000, 1000000, 0, 0, 4) }
+//func BenchmarkWriteOver1Million_5000000(t *testing.B) { runBenchmark(t, 5000000, 1000000, 0, 0, 4) }
+//func BenchmarkWriteOver1Million_10000000(t *testing.B) { runBenchmark(t, 10000000, 1000000, 0, 0, 4) }
 
-func BenchmarkReadOver1Million_10000(t *testing.B)   { runBenchmark(t, 1000000, 0, 10000, 0, 8) }
-func BenchmarkReadOver1Million_100000(t *testing.B)  { runBenchmark(t, 1000000, 0, 100000, 0, 6) }
-func BenchmarkReadOver1Million_1000000(t *testing.B) { runBenchmark(t, 1000000, 0, 1000000, 0, 4) }
+//func BenchmarkReadOver1Million_10000(t *testing.B) { runBenchmark(t, 1000000, 0, 10000, 0,8) }
+//func BenchmarkReadOver1Million_100000(t *testing.B) { runBenchmark(t, 1000000, 0, 100000, 0, 6) }
+//func BenchmarkReadOver1Million_1000000(t *testing.B) { runBenchmark(t, 1000000, 0, 1000000, 0, 4) }
 
-func BenchmarkDeleteOver1Million_10000(t *testing.B)   { runBenchmark(t, 1000000, 0, 0, 10000, 8) }
-func BenchmarkDeleteOver1Million_100000(t *testing.B)  { runBenchmark(t, 1000000, 0, 0, 100000, 6) }
-func BenchmarkDeleteOver1Million_1000000(t *testing.B) { runBenchmark(t, 1000000, 0, 0, 1000000, 4) }
+//func BenchmarkDeleteOver1Million_10000(t *testing.B) { runBenchmark(t, 1000000, 0, 0, 10000,8) }
+//func BenchmarkDeleteOver1Million_100000(t *testing.B) { runBenchmark(t, 1000000, 0, 0, 100000,6) }
+//func BenchmarkDeleteOver1Million_1000000(t *testing.B) { runBenchmark(t, 1000000, 0, 0, 1000000, 4) }
 
-func BenchmarkWriteReadOver1Million_10000(t *testing.B) { runBenchmark(t, 1000000, 10000, 10000, 0, 8) }
-func BenchmarkWriteReadOver1Million_100000(t *testing.B) {
-	runBenchmark(t, 1000000, 100000, 100000, 0, 6)
-}
-func BenchmarkWriteReadOver1Million_1000000(t *testing.B) {
-	runBenchmark(t, 1000000, 1000000, 1000000, 0, 4)
-}
+//func BenchmarkWriteReadOver1Million_10000(t *testing.B) { runBenchmark(t, 1000000, 10000, 10000, 0,8) }
+//func BenchmarkWriteReadOver1Million_100000(t *testing.B) { runBenchmark(t, 1000000, 100000, 100000, 0,6) }
+//func BenchmarkWriteReadOver1Million_1000000(t *testing.B) {runBenchmark(t, 1000000, 1000000, 1000000, 0, 4)}
 
-func BenchmarkWriteReadDeleteOver1Million_10000(t *testing.B) {
-	runBenchmark(t, 1000000, 10000, 10000, 10000, 8)
-}
-func BenchmarkWriteReadDeleteOver1Million_100000(t *testing.B) {
-	runBenchmark(t, 1000000, 100000, 100000, 100000, 6)
-}
-func BenchmarkWriteReadDeleteOver1Million_1000000(t *testing.B) {
-	runBenchmark(t, 1000000, 1000000, 1000000, 1000000, 4)
-}
+//func BenchmarkWriteReadDeleteOver1Million_10000(t *testing.B) { runBenchmark(t, 1000000, 10000, 10000, 10000,8) }
+//func BenchmarkWriteReadDeleteOver1Million_100000(t *testing.B) { runBenchmark(t, 1000000, 100000, 100000, 100000,6) }
+//func BenchmarkWriteReadDeleteOver1Million_1000000(t *testing.B) {runBenchmark(t, 1000000, 1000000, 1000000, 1000000, 4)}
